Alpha 1.6.
Plusieurs petites modifications et petits ajustements.

AJOUT MAJEUR: Réception de toutes les requêtes entrantes via le liber-serveur lorsqu'on ne peut pas créer un serveur local (lorsqu'il est impossible d'ouvrir un port sur un routeur du réseau local). Je ne voulais vraiment pas passer par cette option, mais elle semble incontournable pour que le logiciel puisse fonctionner en tout temps.

AJOUT MAJEUR: La version actuelle du liber-serveur utilise une base de données totalement gérée sur disque via le système de fichiers. Pas de SQL dans la version actuelle. À voir si c'est une bonne option sur le long terme.

+++ src/liber/gui/control/WorkController.java
@@ -254,7 +254,7 @@ public class WorkController implements Controller {
 		}
 	}
 	public void load(Object resource) throws Exception {
-		int tabIndex = resource instanceof Integer ? ((Integer)resource).intValue() : WorkForm.CONTACTS;
+		int tabIndex = resource instanceof Integer ? (Integer) resource : WorkForm.CONTACTS;
 		GUI.current.notifier().setInformer(new WorkInformer());
 		Account account = Libersaurus.current.account();
 		tabs.getSelectionModel().select(tabIndex);
